organized merged files
